refactor: separate concerns in ZenExplorerApp
Refactor ZenExplorerApp.js to follow a more modular architecture:
- Extracted navigation logic into ZenNavigationController.
- Extracted directory rendering and inline rename logic into ZenDirectoryView.
- Extracted Floppy/CD management into ZenDriveManager.
- Extracted context menu building into ZenContextMenuBuilder.
- Extracted keyboard handling into ZenKeyboardHandler.
- Updated README.md to document the new organization.
- Maintained backward compatibility for MenuBarBuilder via proxy methods.
